requirements.txt
# Knowledge Graph Builder 依赖列表
# 核心框架
streamlit==1.32.0

# 大语言模型相关
langchain==0.1.15
langchain-openai==0.0.8

# 文档处理
python-docx==0.8.11
python-dotenv==1.0.0

# 图数据库
neo4j==5.13.0

# 数据处理
pyyaml==6.0.1
json5==0.9.14

# 工具库
requests==2.31.0